Added support for getClassMetadata with the DDc library bundle format
