+++ dwds/test/instances/class_inspection_amd_test.dart
@@ -1,0 +1,35 @@
+// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+@Tags(['daily'])
+@TestOn('vm')
+@Timeout(Duration(minutes: 2))
+library;
+
+import 'package:test/test.dart';
+import 'package:test_common/test_sdk_configuration.dart';
+
+import '../fixtures/context.dart';
+import 'common/class_inspection_common.dart';
+
+void main() {
+  // Enable verbose logging for debugging.
+  final debug = false;
+  final canaryFeatures = false;
+
+  group('Class |', () {
+    final provider = TestSdkConfigurationProvider(
+      verbose: debug,
+    );
+    tearDownAll(provider.dispose);
+    for (final compilationMode in CompilationMode.values) {
+      runTests(
+        provider: provider,
+        compilationMode: compilationMode,
+        canaryFeatures: canaryFeatures,
+        debug: debug,
+      );
+    }
+  });
+}
